+++ src/commands/highlights.js
@@ -195,7 +195,7 @@ class HighLighting extends CommandExecutor {
                         displays.push(`${uid}`)
                     }
                 } catch (err) {
-                    console.error(`獲取用戶資訊錯誤: ${err}`)
+                    console.warn(`獲取用戶資訊錯誤: ${err}`)
                     displays.push(`${uid}`)
                 }
             }else{